requirements.txt
acres==0.3.0
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
configobj==5.0.9
configparser==7.2.0
distro==1.9.0
dotenv==0.9.9
etelemetry==0.3.1
faiss-cpu==1.10.0
filelock==3.18.0
fitz==0.0.1.dev2
Flask==3.1.0
frontend==0.0.3
fsspec==2025.3.1
groq==0.20.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
importlib_resources==6.5.2
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
looseversion==1.3.0
lxml==5.3.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numpy==1.24.4
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pillow==11.1.0
prov==2.0.1
puremagic==1.28
pydantic==2.11.1
pydantic_core==2.33.0
pydot==3.0.4
PyMuPDF==1.25.4
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytils==0.4.1
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rank-bm25==0.2.2
rdflib==6.3.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.0.1
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
sympy==1.13.3
threadpoolctl==3.6.0
tokenizers==0.21.1
tools==0.1.9
torch==2.2.2
tqdm==4.67.1
traits==7.0.2
transformers==4.50.3
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
